requirements.txt
firebase-admin
customtkinter
openpyxl
pandas
pywin32
requests
pillow
google-generativeai
